chore: removes rebase artefact in tealium
